ocr-mosip-integration/requirements.txt
flask==2.3.0
flask-cors==4.0.0
python-dotenv==1.0.0
pytesseract==0.3.10
opencv-python==4.7.0.72
pillow==9.5.0
easyocr==1.7.0
requests==2.31.0
numpy==1.24.0
Werkzeug==2.3.0
qrcode==7.4.2
cbor2==5.4.6
pyzbar==0.1.9
sentence-transformers
scikit-learn
